Add asserts for empty strings.
