+++ utils.js
@@ -15,27 +15,6 @@ const SUPPORTED_KEY_TYPES = [
   'X25519KeyAgreementKey2019'
 ];
 
-/**
-  * Ensures an expected string is not empty.
-  *
-  * @param {object} options - Options to use.
-  * @param {string} options.value - The expected string.
-  * @param {string} options.key - The identifier for the parameter.
-  *
-  * @throws - If the value is not a string or is empty.
-  * @returns {undefined} - It just throws or you are ok.
-*/
-export function notEmptyString({value, key}) {
-  if(typeof value !== 'string') {
-    throw new TypeError(`"${key}" must be a string.`);
-  }
-  if(value.length === 0) {
-    const DataError = new Error(`"${key}" can not be an empty string`);
-    DataError.name = 'DataError';
-    throw DataError;
-  }
-}
-
 export async function delegateCapability(
   {signer, keystoreId, request} = {}) {
   const {
@@ -287,10 +266,4 @@ export function deriveKeystoreId(id) {
     paths[3]; // "<keystore_id>"
 }
 
-export default {
-  notEmptyString,
-  delegateCapability,
-  id,
-  delegate,
-  deriveKeystoreId
-};
+export default {delegateCapability, id, delegate, deriveKeystoreId};
